requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.0.6
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-daterangepicker==1.0.6
django-bootstrap4==22.1
django-bootstrap5==21.3
django-crispy-forms==1.14.0
django-extensions==3.2.0
lazy-object-proxy==1.7.1
Pillow==9.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
